Data.Game

Description

This module contains all data types supporting a game of tic-tac-toe. In particular, it is a game where humans are always first to play.

The GameBoard

data GameBoard

Game boards

A tic-tac-toe game board. The constructor should be treated as private unless within the Game module. Most operations should make use of blankBoard, a GameBoard with no pieces. There is also testBoard which provides a GameBoard for testing only.

Example

Expand

>>> blankBoard
┏━━━┳━━━┳━━━┓
┃ 1 ┃ 2 ┃ 3 ┃
┣━━━╋━━━╋━━━┫
┃ 4 ┃ 5 ┃ 6 ┃
┣━━━╋━━━╋━━━┫
┃ 7 ┃ 8 ┃ 9 ┃
┗━━━┻━━━┻━━━┛
>>> testBoard
┏━━━┳━━━┳━━━┓
┃ X ┃ O ┃ X ┃
┣━━━╋━━━╋━━━┫
┃ 1 ┃ X ┃ O ┃
┣━━━╋━━━╋━━━┫
┃ X ┃ 2 ┃ O ┃
┗━━━┻━━━┻━━━┛

Useful GameBoards

blankBoard :: GameBoard

A completely blank GameBoard.

testBoard :: GameBoard

A GameBoard arbitrarily filled with random pieces for testing only. The actual pieces in testBoard are shown in the documentation for GameBoard.

Operations on GameBoards

countBlanks

Arguments

:: GameBoard
The GameBoard

-> Int
The number of blank positions in the GameBoard

The number of blank positions in a GameBoard

Examples

Expand

>>> countBlanks blankBoard
9
>>> countBlanks testBoard
2

setBlank

Arguments

:: Int
n, i.e. the position to set

-> GameBoard
The GameBoard to set

-> GameBoard
The resulting GameBoard where the nth position is set with the current player's piece

Sets the nth blank position in the GameBoard with the current player's piece. If n is out of bounds, then the GameBoard is returned unmodified.

Examples

Expand

>>> setBlank 9 blankBoard
┏━━━┳━━━┳━━━┓
┃ 1 ┃ 2 ┃ 3 ┃
┣━━━╋━━━╋━━━┫
┃ 4 ┃ 5 ┃ 6 ┃
┣━━━╋━━━╋━━━┫
┃ 7 ┃ 8 ┃ X ┃
┗━━━┻━━━┻━━━┛
ghci> setBlank (-1) blankBoard
┏━━━┳━━━┳━━━┓
┃ 1 ┃ 2 ┃ 3 ┃
┣━━━╋━━━╋━━━┫
┃ 4 ┃ 5 ┃ 6 ┃
┣━━━╋━━━╋━━━┫
┃ 7 ┃ 8 ┃ 9 ┃
┗━━━┻━━━┻━━━┛
ghci> setBlank 2 testBoard
┏━━━┳━━━┳━━━┓
┃ X ┃ O ┃ X ┃
┣━━━╋━━━╋━━━┫
┃ 1 ┃ X ┃ O ┃
┣━━━╋━━━╋━━━┫
┃ X ┃ O ┃ O ┃
┗━━━┻━━━┻━━━┛

Status of games

data GameStatus

The status of the current game

Constructors

Winner
The game has concluded with a winner

Fields

Player

The winning player

Draw
The game has concluded with a draw

Ongoing
The game is still ongoing

Fields

Player

The current player's turn

Obtaining the status of a game

gameStatus

Arguments

:: GameBoard
The GameBoard

-> GameStatus
The status of GameBoard

Obtains the status of the game based on the GameBoard

Examples

Expand

>>> gameStatus blankBoard
Ongoing Human
>>> gameStatus testBoard
Winner Human

Players

data Player

A player in the game.

Constructors

Human
The human

Bot
The bot