Game

Documentation

readPosition :: GameM Int

Repeatedly reads a line from a user until the user enters a valid blank position on the board.

randomPosition :: GameM Int

Generates a random blank position on the board

game :: GameM ()

Plays a game of monadic tic tac toe