Game.Monad

Description

This module defines the main game monad.

Documentation

type RandomSeed = Integer

The seed for random number generation. We use Integers for seeds

type GameState = (GameBoard, RandomSeed)

The GameM monad uses a GameBoard and a RandomSeed as its state.

type GameM = StateT GameState IO

The GameM monad is a StateT monad transformer over the IO monad that uses GameState for its state. Thus, we can use the StateT methods like put, get and evalStateT on it.

getBoard :: GameM GameBoard

Get the current GameBoard

getSeed :: GameM RandomSeed

Get the current RandomSeed

putBoard :: GameBoard -> GameM ()

Put a new GameBoard

putSeed :: RandomSeed -> GameM ()

Put a new RandomSeed

putMsg :: String -> GameM ()

Prints a string to the console

readLine :: GameM String

Reads a line from user input in the console

getGameStatus :: GameM GameStatus

Get the current GameStatus

setAt :: Int -> GameM ()

Sets the nth blank position on the board with the current player's piece

randomInt :: GameM Int

Generates a new random Int

evalStateT :: Monad m => StateT s m a -> s -> m a

Evaluate a state computation with the given initial state and return the final value, discarding the final state.

evalStateT m s = liftM fst (runStateT m s)